+++ kernel/filesystem.h
@@ -36,8 +36,8 @@ int filesystem_get_file_info(file_info* dst, const file_handle* src);
 
 file_stream* filesystem_open_file_handle(const file_handle* f, int mode);
 file_stream* filesystem_open_file(directory_stream* rel, std::string_view path, int mode);
-int filesystem_read_file(void* buf, size_t len, file_stream* f);
-int filesystem_write_file(const void* buf, size_t len, file_stream* f);
+size_t filesystem_read_file(void* buf, size_t len, file_stream* f);
+size_t filesystem_write_file(const void* buf, size_t len, file_stream* f);
 void filesystem_seek_file(file_stream* f, size_t pos);
 size_t filesystem_get_pos(file_stream* f);
 int filesystem_close_file(file_stream* f);
@@ -60,8 +60,9 @@ SYSCALL_HANDLER directory_stream* syscall_open_directory_handle(const file_handl
 SYSCALL_HANDLER int syscall_close_directory(directory_stream* dir);
 SYSCALL_HANDLER file_stream* syscall_open_file_handle(const file_handle* f, int mode);
 SYSCALL_HANDLER file_stream* syscall_open_file(directory_stream* rel, const char* path, size_t path_len, int mode);
-SYSCALL_HANDLER int syscall_read_file(void* dst, size_t len, file_stream* f);
-SYSCALL_HANDLER int syscall_write_file(const void* dst, size_t len, file_stream* f);
+SYSCALL_HANDLER size_t syscall_read_file(void* dst, size_t len, file_stream* f);
+SYSCALL_HANDLER size_t syscall_write_file(const void* dst, size_t len,
+										  file_stream* f);
 SYSCALL_HANDLER int syscall_close_file(file_stream* f);
 SYSCALL_HANDLER int syscall_delete_file(const file_handle* f);
 SYSCALL_HANDLER int syscall_dispose_file_handle(const file_handle* f);
@@ -104,17 +105,17 @@ namespace fs
 	namespace {
 		class stream_base {
 		public:
-			int read(void* dst, size_t len)
+			size_t read(void* dst, size_t len)
 			{
 				return filesystem_read_file(dst, len, get_ptr());
 			}
 
-			int write(const void* dst, size_t len)
+			size_t write(const void* dst, size_t len)
 			{
 				return filesystem_write_file(dst, len, get_ptr());
 			}
 
-			int write(const uint8_t dst)
+			size_t write(const uint8_t dst)
 			{
 				return filesystem_write_file(&dst, 1, get_ptr());
 			}
